consume tests
